# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Josafá Hair & Prótese Capilar
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface Appointment {
  id: number;
  name: string;
  phone: string;
  email: string;
  serviceId: number;
  serviceName: string;
  preferredDate: string;
  preferredTime: string;
  notes?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}

export interface CreateAppointmentBody {
  name: string;
  phone: string;
  email: string;
  serviceId: number;
  preferredDate: string;
  preferredTime: string;
  notes?: string | null;
}

export type UpdateAppointmentStatusBodyStatus =
  (typeof UpdateAppointmentStatusBodyStatus)[keyof typeof UpdateAppointmentStatusBodyStatus];

export const UpdateAppointmentStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface UpdateAppointmentStatusBody {
  status: UpdateAppointmentStatusBodyStatus;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  duration: string;
  category: string;
}

export interface AppointmentStats {
  total: number;
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
}

export interface BlockedSlot {
  id: number;
  date: string;
  time?: string | null;
  reason?: string | null;
  createdAt: string;
}

export interface CreateBlockedSlotBody {
  date: string;
  time?: string | null;
  reason?: string | null;
}

export interface ErrorResponse {
  error: string;
}