# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Josafá Hair & Prótese Capilar
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: appointments
    description: Appointment booking operations
  - name: services
    description: Services offered by the salon
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  /blocked-slots:
    get:
      operationId: listBlockedSlots
      tags: [appointments]
      summary: List all blocked slots
      responses:
        "200":
          description: List of blocked slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedSlot"
    post:
      operationId: createBlockedSlot
      tags: [appointments]
      summary: Create a blocked slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlockedSlotBody"
      responses:
        "201":
          description: Blocked slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedSlot"

  /blocked-slots/{id}:
    delete:
      operationId: deleteBlockedSlot
      tags: [appointments]
      summary: Delete a blocked slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /appointments/stats:
    get:
      operationId: getAppointmentStats
      tags: [appointments]
      summary: Get appointment statistics
      responses:
        "200":
          description: Appointment statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Appointment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        preferredDate:
          type: string
        preferredTime:
          type: string
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - email
        - serviceId
        - serviceName
        - preferredDate
        - preferredTime
        - status
        - createdAt

    CreateAppointmentBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        serviceId:
          type: integer
        preferredDate:
          type: string
        preferredTime:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - name
        - phone
        - email
        - serviceId
        - preferredDate
        - preferredTime

    UpdateAppointmentStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
      required:
        - status

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        duration:
          type: string
        category:
          type: string
      required:
        - id
        - name
        - description
        - duration
        - category

    AppointmentStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        confirmed:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
      required:
        - total
        - pending
        - confirmed
        - completed
        - cancelled

    BlockedSlot:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        time:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - date
        - createdAt

    CreateBlockedSlotBody:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
      required:
        - date

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error