# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Josafá Hair & Prótese Capilar
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all appointments
 */
export const ListAppointmentsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  serviceId: zod.number(),
  serviceName: zod.string(),
  preferredDate: zod.string(),
  preferredTime: zod.string(),
  notes: zod.string().nullish(),
  status: zod.enum(["pending", "confirmed", "cancelled", "completed"]),
  createdAt: zod.string(),
});
export const ListAppointmentsResponse = zod.array(ListAppointmentsResponseItem);

/**
 * @summary Create a new appointment
 */
export const CreateAppointmentBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  serviceId: zod.number(),
  preferredDate: zod.string(),
  preferredTime: zod.string(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get appointment by ID
 */
export const GetAppointmentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAppointmentResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  serviceId: zod.number(),
  serviceName: zod.string(),
  preferredDate: zod.string(),
  preferredTime: zod.string(),
  notes: zod.string().nullish(),
  status: zod.enum(["pending", "confirmed", "cancelled", "completed"]),
  createdAt: zod.string(),
});

/**
 * @summary Update appointment status
 */
export const UpdateAppointmentStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateAppointmentStatusBody = zod.object({
  status: zod.enum(["pending", "confirmed", "cancelled", "completed"]),
});

export const UpdateAppointmentStatusResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  serviceId: zod.number(),
  serviceName: zod.string(),
  preferredDate: zod.string(),
  preferredTime: zod.string(),
  notes: zod.string().nullish(),
  status: zod.enum(["pending", "confirmed", "cancelled", "completed"]),
  createdAt: zod.string(),
});

/**
 * @summary List all blocked slots
 */
export const ListBlockedSlotsResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  time: zod.string().nullish(),
  reason: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListBlockedSlotsResponse = zod.array(ListBlockedSlotsResponseItem);

/**
 * @summary Create a blocked slot
 */
export const CreateBlockedSlotBody = zod.object({
  date: zod.string(),
  time: zod.string().nullish(),
  reason: zod.string().nullish(),
});

/**
 * @summary Delete a blocked slot
 */
export const DeleteBlockedSlotParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all services
 */
export const ListServicesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  duration: zod.string(),
  category: zod.string(),
});
export const ListServicesResponse = zod.array(ListServicesResponseItem);

/**
 * @summary Get appointment statistics
 */
export const GetAppointmentStatsResponse = zod.object({
  total: zod.number(),
  pending: zod.number(),
  confirmed: zod.number(),
  completed: zod.number(),
  cancelled: zod.number(),
});