# lib/api-zod/src/generated/types/blockedSlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Josafá Hair & Prótese Capilar
 * OpenAPI spec version: 0.1.0
 */

export interface BlockedSlot {
  id: number;
  date: string;
  time?: string | null;
  reason?: string | null;
  createdAt: string;
}